collectionをuiStateに置換
