+++ core-design/src/main/java/com/kabos/topicker/core/design/component/BubbleLoading.kt
@@ -1,0 +1,22 @@
+package com.kabos.topicker.core.design.component
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Modifier
+import com.airbnb.lottie.compose.*
+import com.kabos.topicker.core.design.R
+
+
+@Composable
+fun BubbleLoading(modifier: Modifier = Modifier) {
+    val composition by rememberLottieComposition(LottieCompositionSpec.RawRes(R.raw.loading_bubbles))
+    val progress by animateLottieCompositionAsState(
+        composition = composition,
+        iterations = LottieConstants.IterateForever,
+    )
+    LottieAnimation(
+        composition = composition,
+        progress = { progress },
+        modifier = modifier
+    )
+}
